Content Timeline - Class migration
